control-chart-snapshot-bot/requirements.txt

requests==2.32.3
pandas==2.2.2
matplotlib==3.9.0
python-dotenv==1.0.1
